create: sendCodefile, and call sendCodeToEmail in signup.ts
